Fixed scaled rendering for index based painting
When rendering was switched to be index based (instead of `GetPixel()/SetPixel()` based), a mistake was made where the initial color bytes were written to the display buffer but the scaling/replicated bytes as written to the source/pixel buffer. This caused the display to never get the scaling bytes, and therefore the display was only scaled in one direction.

This fixes that bug and has scaled rendering working again.  Some slight refactoring, naming, and small performance improvements was done to make `ShowFastMode()` performance close to non-scaled rendering (only 10ms slower for scale of 2).