Removed display rotation from engine
Currently, when we draw a sprite we go transform every pixel's x and y coordinates based on the display rotation.

Not only does this add extra complexity and CPU overhead for each sprite that needs to be rendered, it also limits rendering optimizations that can be done.

Instead, we can rely on individual display drivers managing the rotation.  This allows us to manage the byte buffer and sprites without rotations, and allow the display driver to properly handle the rotation.

This won't be functional right away, but requires a Meadow PR to land and propagate out to nuget (https://github.com/WildernessLabs/Meadow.Foundation/pull/528).