Don't rotate the display for project lab
Previously, we moved rotation out of the engine and instead used the St7789's `SetRotation()` method to perform this.

However, until a Meadow labs bug is merged and released in a nuget, this is buggy.  So to prevent this from disrupting development efforts before then the display should have no rotation.